refactor(orchestrator): move ResultToGetTargetGraphResponse out of internal/idmap
ResultToGetTargetGraphResponse only had one real consumer, orchestrator's
GetTargetGraph; move it there instead of a shared internal package, and
have example/cmd/query-bench call it directly off orchestrator.

Rename internal/targetgraph to internal/idmap: with the conversion
function gone, the package only holds NameIDMapper and the tangopb
metadata-map chunking helpers, so idmap reflects its actual contents.
