+++ controller/BUILD.bazel
@@ -18,7 +18,7 @@ go_library(
         "//core/common",
         "//core/storage",
         "//internal/cachekey",
-        "//internal/targetgraph",
+        "//internal/idmap",
         "//orchestrator",
         "//tangopb",
         "@com_github_uber_go_tally//:tally",
@@ -42,7 +42,7 @@ go_test(
         "//core/common",
         "//core/storage",
         "//core/storage/storagemock",
-        "//internal/targetgraph",
+        "//internal/idmap",
         "//orchestrator/orchestratormock",
         "//tangopb",
         "//tangopb/tangopbmock",
