+++ controller/testhelper_test.go
@@ -18,7 +18,7 @@ import (
 	"context"
 
 	"github.com/uber-go/tally"
-	"github.com/uber/tango/internal/targetgraph"
+	"github.com/uber/tango/internal/idmap"
 	"go.uber.org/zap"
 )
 
@@ -26,9 +26,9 @@ func newTestController(logger *zap.Logger) *controller {
 	return &controller{
 		logger:                 logger,
 		scope:                  tally.NoopScope,
-		targetChunkSize:        targetgraph.DefaultTargetChunkSize,
-		changedTargetChunkSize: defaultChangedTargetChunkSize,
-		metadataMapChunkSize:   targetgraph.DefaultMetadataMapChunkSize,
+		targetChunkSize:        idmap.DefaultTargetChunkSize,
+		changedTargetChunkSize: _defaultChangedTargetChunkSize,
+		metadataMapChunkSize:   idmap.DefaultMetadataMapChunkSize,
 		totalDurationBuckets:   _totalDurationBuckets,
 		appCtx:                 context.Background(),
 	}
