+++ internal/cachekey/cachekey_test.go
@@ -138,7 +138,7 @@ func TestGetReqURLsHash(t *testing.T) {
 	for _, tt := range tests {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
-			assert.Equal(t, tt.want, GetReqURLsHash(tt.in))
+			assert.Equal(t, tt.want, getReqURLsHash(tt.in))
 		})
 	}
 }
