--- internal/targetgraph/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "targetgraph",
-    srcs = [
-        "nameidmapper.go",
-        "targetgraph.go",
-    ],
-    importpath = "github.com/uber/tango/internal/targetgraph",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//core/targethasher",
-        "//tangopb",
-        "@com_github_bazelbuild_buildtools//build_proto",
-    ],
-)
-
-go_test(
-    name = "targetgraph_test",
-    srcs = [
-        "nameidmapper_test.go",
-        "targetgraph_test.go",
-    ],
-    embed = [":targetgraph"],
-    deps = [
-        "//core/targethasher",
-        "//tangopb",
-        "@com_github_stretchr_testify//assert",
-        "@com_github_stretchr_testify//require",
-    ],
-)
